fix(ui): update segmented control imports
