+++ src/ui/components/legacy/index.ts
@@ -1,10 +1,5 @@
-export { Button } from './Button';
-export { InputField } from './InputField';
-export { Checkbox } from './Checkbox';
 export { Select, SelectOption } from './Select';
 export { Text } from './Text';
 export { Icon } from './Icon';
 export { Paragraph } from './Paragraph';
 export { FormGroup } from './FormGroup';
-export { Panel } from './Panel';
-export { Section } from './Section';
