+++ src/ui/pages/ArrangeTab.tsx
@@ -1,14 +1,6 @@
 import React from 'react';
-import {
-  Button,
-  Checkbox,
-  InputField,
-  Icon,
-  Select,
-  SelectOption,
-  Text,
-} from '../components/legacy';
-import { Panel } from '../components/legacy';
+import { Button, Checkbox, InputField, Panel } from '../components';
+import { Icon, Select, SelectOption, Text } from '../components/legacy';
 import { TabGrid } from '../components/TabGrid';
 import { applyGridLayout, GridOptions } from '../../board/grid-tools';
 import { applySpacingLayout, SpacingOptions } from '../../board/spacing-tools';
