+++ src/ui/pages/DiagramsTab.tsx
@@ -5,7 +5,7 @@ import { LayoutEngineTab } from './LayoutEngineTab';
 import { TabPanel } from '../components/TabPanel';
 import type { TabTuple } from './tab-definitions';
 import { Tabs } from '@mirohq/design-system';
-import { Panel } from '../components/legacy';
+import { Panel } from '../components';
 
 /**
  * Parent tab hosting diagram-related tools via nested navigation.
