+++ src/ui/pages/StyleTab.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { Button, Icon, InputField, Text } from '../components/legacy';
+import { Button, InputField, Panel } from '../components';
+import { Icon, Text } from '../components/legacy';
 import { tweakFillColor, extractFillColor } from '../../board/style-tools';
 import { applyStylePreset, presetStyle } from '../../board/format-tools';
 import { STYLE_PRESET_NAMES, stylePresets } from '../style-presets';
@@ -10,7 +11,6 @@ import { TabPanel } from '../components/TabPanel';
 import { TabGrid } from '../components/TabGrid';
 import type { TabTuple } from './tab-definitions';
 import { Heading } from '@mirohq/design-system';
-import { Panel } from '../components/legacy';
 
 /** Adjusts the fill colour of selected widgets. */
 export const StyleTab: React.FC = () => {
