+++ tests/button-size.test.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { Button } from '../src/ui/components/legacy/Button';
+import { Button } from '../src/ui/components/Button';
 
 it('defaults to medium size for primary', () => {
   const { getByRole } = render(<Button variant='primary'>Ok</Button>);
