+++ tests/inputfield.test.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { InputField } from '../src/ui/components/legacy/InputField';
+import { InputField } from '../src/ui/components/InputField';
 
 test('renders label and input', () => {
   render(
