+++ tests/panel-section.test.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { Panel, Section } from '../src/ui/components/legacy';
+import { Panel, Section } from '../src/ui/components';
 import { tokens } from '../src/ui/tokens';
 
 describe('Panel', () => {
